backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
dj-rest-auth==7.0.0
Django==4.2.6
django-allauth==0.61.1
django-cors-headers==4.2.0
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.4
oauthlib==3.2.2
pillow==11.0.0
psycopg2-binary==2.9.10
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
